+++ axibank/src/main/java/com/app/axibank/service/CustomerService.java
@@ -1,10 +1,15 @@
 package com.app.axibank.service;
 
+import com.app.axibank.mapper.CustomerMapper;
 import com.app.axibank.model.customer.Customer;
+import com.app.axibank.model.customer.CustomerRequestDTO;
+import com.app.axibank.model.customer.CustomerResponseDTO;
 import com.app.axibank.repository.CustomerRepository;
+import jakarta.transaction.Transactional;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
-import org.springframework.transaction.annotation.Transactional;
+
+import java.util.Optional;
 
 @Service
 @RequiredArgsConstructor
@@ -13,16 +18,45 @@ public class CustomerService {
     private final CustomerRepository customerRepository;
 
     @Transactional
-    public Customer create(Customer customer) {
-        return customerRepository.save(customer);
+    public CustomerResponseDTO create(CustomerRequestDTO request) {
+        if (customerRepository.existsByCpf(request.getCpf())) {
+            throw new IllegalArgumentException("CPF já cadastrado.");
+        }
+
+        if (customerRepository.existsByEmail(request.getEmail())) {
+            throw new IllegalArgumentException("Email já cadastrado.");
+        }
+
+        Customer customer = CustomerMapper.toEntity(request);
+        Customer saved = customerRepository.save(customer);
+        return CustomerMapper.toDTO(saved);
+    }
+
+    public CustomerResponseDTO getById(Long id) {
+        Customer customer = customerRepository.findById(id)
+                .orElseThrow(() -> new IllegalArgumentException("Cliente não encontrado."));
+        return CustomerMapper.toDTO(customer);
     }
 
     @Transactional
-    public Customer updateCustomer(Customer customer) {
-        return customerRepository.save(customer);
+    public CustomerResponseDTO update(Long id, CustomerRequestDTO request) {
+        Customer customer = customerRepository.findById(id)
+                .orElseThrow(() -> new IllegalArgumentException("Cliente não encontrado."));
+
+        customer.setName(request.getName());
+        customer.setCpf(request.getCpf());
+        customer.setEmail(request.getEmail());
+        customer.setTelephone(request.getTelephone());
+
+        Customer updated = customerRepository.save(customer);
+        return CustomerMapper.toDTO(updated);
     }
+
     @Transactional
-    public void deleteCustomer(Long id) {
+    public void delete(Long id) {
+        if (!customerRepository.existsById(id)) {
+            throw new IllegalArgumentException("Cliente não encontrado.");
+        }
         customerRepository.deleteById(id);
     }
 }
